Simple Game
Added pictures to create "player" prefabs. Added scripts to create simple game which spawns 3x3D markers on a map. Player 1 can be controlled by the buttons in the top left corner and coins are randomly spawned near the default spawn location. The player can move and hit the coins collecting them and increasing their score which is displayed on the screen.